Move sync_channelmetadata into utils/annotation.py
